refactored connectivity_models
