--- package/cloudshell/cp/aws/models/connectivity_models.py
@@ -1,22 +1,0 @@
-# todo: move to network_actions_model.py
-
-class ConnectivityActionResult(object):
-    def __init__(self):
-        self.actionId = ''
-        self.success = True
-        self.infoMessage = ''
-        self.errorMessage = ''
-
-
-class PrepareNetworkActionResult(ConnectivityActionResult):
-    def __init__(self):
-        ConnectivityActionResult.__init__(self)
-        self.vpcId = ''
-        self.securityGroupId = ''
-        self.type = 'PrepareNetwork'
-
-
-class PrepareSubnetActionResult(ConnectivityActionResult):
-    def __init__(self):
-        ConnectivityActionResult.__init__(self)
-        self.subnetId = ''
